[OSE] Youtube Video Playback sound is heard , When volume is in mute.
:Release Notes:
Solves the issue of volume is heard when youtube
playback is paused - muted - played.

:Detailed Notes:
solves the issue of audiod not sending volume set command
to pulseaudio when volume requested is zero and no on active streams
are present in pulseaudio.

:Testing Performed:
Compiled AudioD locally and verified on latest OSE #25 build.
Use cases verified are as below.
1) Open browser and open wwww.youtube.com
2) Play any video - Audio heard
3) Press pause button in the youtube video window
4) Press mute button in the youtube video window
5) Press play button
Bug Result    - Audio heard
Actual Result - Audio not heard

:QA Notes:
NA

:Issues Addressed:
[PLAT-65265] [webOS OSE] : Youtube Video Playback sound is heard ,
             When volume is in mute.

Change-Id: I26301e583749057611f4b6803faaf07d98cf5a65