Implement sink status notification in audiod
:Release Notes:
Implement sink status notification in audiod to notify the
change in sink status to modules

:Detailed Notes:
Change in sink status information is required by modules to
apply the volume policy for audio streams.
This patch implements sink status notification that modules
can subscribe to get the sink status, when a stream is opened
or closed.

:Testing Performed:
Tested on RPI4 rebased on #220 ose official binary
Test cases:
1. Check if audiod service is running : ps -ef | grep audiod
2. QWA-4099
Test Results are attached to PLAT-112775

:QA Notes:
NA

:Issues Addressed:
[PLAT-112775] Implement sink status notification in audiod

Change-Id: I849286bb147e07fb4a636ba7594b6b2fc42006cd