Fix yocto version update in Audiod
:Release Notes:
Implement yocto version update in Audiod

:Detailed Notes:
Yocto build is updated to version 3.1 and in new version of yocto
build, multithreading and static libraries are not required.
This patch removes the inclusion of static libraries.

:Testing Performed:
Tested on RPI4 rebased on #234 ose official binary
Test cases: QWA-4099
Test Results are attached to PLAT-116443

:QA Notes:
NA

:Issues Addressed:
[PLAT-116443] Fix yocto version update in Audiod

Change-Id: I443683cefbc72d18a934df225b99bd4b48ed1546